move rnafusion model dir
