+++ cg/models/rnafusion/rnafusion.py
@@ -6,7 +6,7 @@ from pydantic.v1 import BaseModel
 from cg.constants.constants import Strandedness
 from cg.models.analysis import AnalysisModel
 from cg.models.nextflow.sample import NextflowSample
-from cg.models.workflow.nf_analysis import PipelineParameters
+from cg.models.nf_analysis import PipelineParameters
 
 
 class RnafusionQCMetrics(BaseModel):
